display the old price

+++ FilRouge-app/database/migrations/2024_04_16_151139_create_products_table.php
@@ -11,20 +11,22 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('products', function (Blueprint $table) {
-            $table->id();
-            $table->string('title');
-            $table->decimal('price', 8, 2);
-            $table->string('product_nbr');
-            $table->text('description');
-            $table->string('image');
-            $table->bigInteger('category');
-            $table->softDeletes();
-            $table->timestamps();
+        if (!Schema::hasTable('products')) {
+            Schema::create('products', function (Blueprint $table) {
+                $table->id();
+                $table->string('title');
+                $table->decimal('price', 8, 2);
+                $table->string('product_nbr');
+                $table->text('description');
+                $table->string('image');
+                $table->bigInteger('category');
+                $table->softDeletes();
+                $table->timestamps();
 
-            $table->foreign('category')->references('id')->on('categories')->onDelete('cascade');
+                $table->foreign('category')->references('id')->on('categories')->onDelete('cascade');
 
-        });
+            });
+        }
     }
 
     /**
